docs: add description for motheds in gpd_helper
